refactor: splitted sources into several files
